requirements.txt
flask==2.3.2
pillow==10.0.0